+++ src/Engine/ParticleSystem/Behaviours/Inits/SquarePosInit.cpp
@@ -2,7 +2,7 @@
 
 namespace ParticleSystem
 {
-	SquarePosInit::SquarePosInit(int p_minx, int p_maxx, int p_miny, int p_maxy ) noexcept
+	SquarePosInit::SquarePosInit(float p_minx, float p_maxx, float p_miny, float p_maxy ) noexcept
 	: _minx { p_minx }, _maxx { p_maxx }, _miny { p_miny }, _maxy { p_maxy }
 	{
         _xDiff = p_maxx - p_minx;
@@ -15,8 +15,8 @@ namespace ParticleSystem
 	void
 	SquarePosInit::init(ParticleSlotmap& p_particles, IdPair& p_particleKey) const noexcept
 	{
-        int randX = rand()%_xDiff - _minx;
-        int randY = rand()%_yDiff - _miny;
+        float randX = rand()%_xDiff - _minx;
+        float randY = rand()%_yDiff - _miny;
 
 		auto& particlePos = p_particles.getItem<Position>(p_particleKey);
